+++ core-client/src/main/scala/de/welt/contentapi/core/client/CoreModule.scala
@@ -2,9 +2,7 @@ package de.welt.contentapi.core.client
 
 import com.google.inject.AbstractModule
 import com.kenshoo.play.metrics.{Metrics, MetricsImpl}
-import de.welt.contentapi.core.client.repository.{PapyrusRepository, PapyrusRepositoryImpl}
 import de.welt.contentapi.core.client.services.contentapi._
-import de.welt.contentapi.core.client.services.curation.{CurationService, CurationServiceImpl}
 import de.welt.contentapi.core.client.services.s3.{S3Client, S3ClientImpl}
 
 class CoreModule extends AbstractModule {
@@ -15,7 +13,5 @@ class CoreModule extends AbstractModule {
     bind(classOf[S3Client]).to(classOf[S3ClientImpl])
     bind(classOf[ContentService]).to(classOf[ContentServiceImpl])
     bind(classOf[ContentSearchService]).to(classOf[ContentSearchServiceImpl])
-    bind(classOf[PapyrusRepository]).to(classOf[PapyrusRepositoryImpl])
-    bind(classOf[CurationService]).to(classOf[CurationServiceImpl])
   }
 }