+++ pressed/src/main/scala/de/welt/contentapi/pressed/models/legacy/ApiEnrichedTeaser.scala
@@ -1,7 +1,0 @@
-package de.welt.contentapi.pressed.models.legacy
-
-import de.welt.contentapi.core.models.ApiContent
-import de.welt.contentapi.pressed.models.ApiChannel
-
-case class ApiEnrichedTeaser(content: Option[ApiContent] = None,
-                             channel: Option[ApiChannel] = None)